refac : 스크랩 페이지 hook 로직 분리
